requirements.txt
httpx>=0.27
tenacity>=8.2
python-dotenv>=1.0
mcp>=1.0
